Election dates working - sort off- clean data not being created cleanly
