+++ data/data_loader.py
@@ -1,7 +1,7 @@
 import streamlit as st
 from data.datasetupandclean import load_raw_data
 from data.clean_and_enhance import load_cleaned_data
-from data.load_donor_regent_lists import load_donorList_data, load_regulated_entity_data, load_entity_summary_data
+from data.load_donor_regent_lists import load_donorList_data, load_regulated_entity_data  # load_entity_summary_data
 from utils.logger import log_function_call
 
 
@@ -18,17 +18,17 @@ def get_raw_data():
         processeddatafilepath="imported_raw_fname")
 
 
-@log_function_call
-@st.cache_data
-def get_party_summary_data():
-    return load_entity_summary_data(        
-        main_file="raw_data",
-        cleaned_file="party_summary_fname",
-        datafile=None,
-        output_csv=True,
-        streamlitrun=True,
-        originaldatafilepath="cleaned_data_fname",
-        cleaneddatafilepath="party_summary_fname")
+# @log_function_call
+# @st.cache_data
+# def get_party_summary_data():
+#     return load_entity_summary_data(        
+#         main_file="raw_data",
+#         cleaned_file="party_summary_fname",
+#         datafile=None,
+#         output_csv=True,
+#         streamlitrun=True,
+#         originaldatafilepath="cleaned_data_fname",
+#         cleaneddatafilepath="party_summary_fname")
 
 
 @log_function_call
@@ -76,17 +76,17 @@ def firstload():
     #     create_thresholds()
 
     # Load and cache data correctly
-
-    @log_function_call
-    def load_data_to_session(key, loader_function):
-        if key not in st.session_state:
-            st.session_state[key] = loader_function()
-            return key
-        else:
-            return key
-
     load_data_to_session("raw_data", get_raw_data)
     load_data_to_session("data_clean", get_cleaned_data)
-    load_data_to_session("data_party_sum", get_party_summary_data)
+    # load_data_to_session("data_party_sum", get_party_summary_data)
     load_data_to_session("data_donor", get_donor_data)
     load_data_to_session("data_regentity", get_regentity_data)
+
+
+@log_function_call
+def load_data_to_session(key, loader_function):
+    if key not in st.session_state:
+        st.session_state[key] = loader_function()
+        return key
+    else:
+        return key
