+++ reference_files/elections.csv
@@ -1,0 +1,38 @@
+type,name,dissolution,election
+General Election,1929,1929-05-10,1929-05-30
+General Election,1931,1931-10-07,1931-10-27
+General Election,1935,1935-10-25,1935-11-14
+General Election,1945,1945-06-15,1945-07-05
+General Election,1950,1950-02-03,1950-02-23
+General Election,1951,1951-10-05,1951-10-25
+General Election,1955,1955-05-06,1955-05-26
+General Election,1959,1959-09-18,1959-10-08
+General Election,1964,1964-09-25,1964-10-15
+General Election,1966,1966-03-10,1966-03-31
+General Election,1970,1970-05-29,1970-06-18
+Referendum,Northern Ireland border poll 1973-NI,1972-11-01,1973-03-08
+General Election,1974 (Feb),1974-02-08,1974-02-28
+General Election,1974 (Oct),1974-09-20,1974-10-10
+Referendum,EC Membership 1975,1975-02-26,1975-06-05
+Referendum,Scottish Devolution 1979-Scotland,1978-07-31,1979-03-01
+Referendum,Welsh Devolution 1979-Wales,1978-07-31,1979-03-01
+General Election,1979,1979-04-07,1979-05-03
+General Election,1983,1983-05-13,1983-06-09
+General Election,1987,1987-05-18,1987-06-11
+General Election,1992,1992-03-16,1992-04-09
+General Election,1997,1997-04-08,1997-05-01
+Referendum,Scottish Devolution 1997-Scotland,1997-06-19,1997-09-11
+Referendum,Welsh Devolution 1997-Wales,1997-06-19,1997-09-11
+Referendum,Greater London Authority 1998-London,1998-06-30,1998-09-07
+Referendum,NI Good Friday Agreement 1998-NI,1998-06-30,1998-09-07
+General Election,2001,2001-05-14,2001-06-07
+Referendum,North East Assembly 2004-North East,2004-06-17,2004-11-04
+General Election,2005,2005-04-11,2005-05-05
+General Election,2010,2010-04-12,2010-05-06
+Referendum,AV Referendum 2011,2011-02-16,2011-05-05
+Referendum,Scottish Independence 2014-Scotland,2013-11-26,2014-09-18
+General Election,2015,2015-03-30,2015-05-07
+Referendum,EU Membership 2016,2015-12-17,2016-06-23
+General Election,2017,2017-05-03,2017-06-08
+General Election,2019,2019-11-06,2019-12-12
+General Election,2024,2024-05-30,2024-07-04
