--- reference_files/potential_regentiity_duplicates.csv
@@ -1,108 +1,0 @@
-RegulatedEntityId,Potential Duplicates
-1444,[1345]
-1345,"[1444, 2005]"
-2423,[1893]
-1893,[2423]
-1028,[1361]
-1361,[1028]
-1622,"[1624, 1799]"
-1624,[1622]
-1799,[1622]
-1371,"[1611, 1301]"
-1301,"[1312, 1264, 1611, 1371]"
-1611,"[1312, 1371, 1301]"
-9050,[4188]
-4188,[9050]
-1079,[1311]
-1311,"[1274, 1079]"
-1334,[1228]
-1228,[1334]
-2450,[2223]
-2223,[2450]
-1274,"[1182, 1311]"
-1182,"[1137, 1274]"
-1825,[1173]
-1173,"[1825, 1575]"
-9153,[6308]
-6308,[9153]
-1284,[6778]
-6778,"[1284, 2039]"
-1796,"[1632, 1528]"
-1632,"[1156, 1796, 1573, 1836, 1528]"
-1528,"[1632, 1796]"
-1778,[1630]
-1630,[1778]
-1591,[1900]
-1900,[1591]
-6591,[5269]
-5269,[6591]
-2005,[1345]
-1828,[1494]
-1494,[1828]
-2039,[6778]
-2587,[1357]
-1357,"[2081, 2587]"
-2061,[1850]
-1850,[2061]
-1050,[1926]
-1926,[1050]
-2081,[1357]
-1288,[1346]
-1346,[1288]
-1895,[2578]
-2578,[1895]
-1575,[1173]
-2796,[1795]
-1795,[2796]
-4003,[916]
-916,[4003]
-4212,[4184]
-4184,[4212]
-1473,[2148]
-2148,[1473]
-1827,[1479]
-1479,[1827]
-1914,"[2130, 1615]"
-1615,"[1914, 2130]"
-2130,"[1914, 1615]"
-1916,[1062]
-1062,[1916]
-2855,[1819]
-1819,[2855]
-1802,[1364]
-1364,[1802]
-1985,[1609]
-1609,[1985]
-1650,[1176]
-1176,[1650]
-2040,[1776]
-1776,[2040]
-1489,[1807]
-1807,[1489]
-1572,[1188]
-1188,[1572]
-1681,[1038]
-1038,[1681]
-2002,[1511]
-1511,[2002]
-1422,[1440]
-1440,[1422]
-1341,[1143]
-1143,[1341]
-1818,[1457]
-1457,[1818]
-1836,[1632]
-1156,[1632]
-1312,"[1264, 1611, 1301]"
-1573,"[1632, 1087]"
-1087,[1573]
-1388,[1280]
-1280,[1388]
-1264,"[1312, 1301]"
-1137,[1182]
-1058,[1031]
-1031,[1058]
-1108,[1077]
-1077,[1108]
-2657,[2489]
-2489,[2657]
