--- utils/decorators.py
@@ -1,2 +1,0 @@
-import utils.logger as logger
-
