+++ utils/global_variables.py
@@ -24,7 +24,7 @@ import os
 def initialize_session_state():
     BASE_DIR = Path(os.getcwd())
     logger.info(f"BASE_DIR: {BASE_DIR}")
-    
+
     # Set globel env variables
     os.environ["LOG_LEVEL"] = config.LOG_LEVEL
     init_state_var("LOG_LEVEL", config.LOG_LEVEL)
