+++ utils/password_hashing.py
@@ -1,5 +1,5 @@
 import bcrypt
 
-password = "Gr3yW0rmD4nc3r"
+password = "J3rryTh3B3rry"
 hashed = bcrypt.hashpw(password.encode(), bcrypt.gensalt())
 print(hashed.decode())  # Copy this and update `admin_credentials.json`
